filter propertyId config

+++ src/main/java/edf/xmlparser/job/XmlParserJob.java
@@ -26,9 +26,11 @@ public class XmlParserJob implements Job {
     public void execute(JobExecutionContext jobExecutionContext) {
         SDIFileInsertProcessor processor = new SDIFileInsertProcessor();
         Long start = System.currentTimeMillis();
-        ZipTools.unzipFolder(PropertyUtil.getPropValue(PropsStr.WorkPath));
+        boolean isSuccessUnzipGzFiles = ZipTools.unzipFolder(PropertyUtil.getPropValue(PropsStr.WorkPath));
         //FileUtils.moveGzFiles(PropertyUtil.getPropValue(PropsStr.WorkPath), PropertyUtil.getPropValue(PropsStr.GzFileAchievePath));
-        FileUtils.deleteGzFiles(PropertyUtil.getPropValue(PropsStr.WorkPath));
+        if (isSuccessUnzipGzFiles) {
+            FileUtils.deleteGzFiles(PropertyUtil.getPropValue(PropsStr.WorkPath));
+        }
         List<File> files = FileUtils.getLocalAbsFiles(PropertyUtil.getPropValue(PropsStr.WorkPath));
         processor.process(files, PropertyUtil.getPropValue(PropsStr.FileAchievePath));
         Long end = System.currentTimeMillis();
